重构redis模块，增加 lettuce 和 redisson 的支持，默认是 jedis
