+++ src/main/java/io/jboot/component/redis/JbootRedisConfig.java
@@ -5,7 +5,7 @@
  * you may not use this file except in compliance with the License.
  * You may obtain a copy of the License at
  * <p>
- *  http://www.apache.org/licenses/LICENSE-2.0
+ * http://www.apache.org/licenses/LICENSE-2.0
  * <p>
  * Unless required by applicable law or agreed to in writing, software
  * distributed under the License is distributed on an "AS IS" BASIS,
@@ -25,6 +25,10 @@ import java.util.Set;
 @PropertyConfig(prefix = "jboot.redis")
 public class JbootRedisConfig {
 
+    public static final String TYPE_JEDIS = "jedis";
+    public static final String TYPE_REDISSON = "redisson";
+    public static final String TYPE_LETTUCE = "lettuce";
+
     private String host;
     private Integer port = 6379;
     private Integer timeout = 2000;
@@ -39,6 +43,7 @@ public class JbootRedisConfig {
     private Long timeBetweenEvictionRunsMillis;
     private Integer numTestsPerEvictionRun;
     private Integer maxAttempts;
+    private String type = TYPE_JEDIS;
 
 
     public String getHost() {
@@ -163,6 +168,14 @@ public class JbootRedisConfig {
 
     public boolean isClusterConfig() {
         return isConfigOk() && host.contains(",");
+    }
+
+    public String getType() {
+        return type;
+    }
+
+    public void setType(String type) {
+        this.type = type;
     }
 
     public Set<HostAndPort> getHostAndPorts() {
